update style function
